cast to Float

+++ scripts/02_write_SEI_to_drive.js
@@ -130,7 +130,8 @@ Export.image.toDrive({
 var bandsForOutput = function(image) {
   var out = image
     .select(['Q1raw', 'Q2raw', 'Q3raw', 'Q4raw', 'Q5raw',"Q5s", 'Q5sc3'])
-    .rename(['Q1raw', 'Q2raw', 'Q3raw', 'Q4raw', 'Q5raw',"SEI", 'sc3']);
+    .rename(['Q1raw', 'Q2raw', 'Q3raw', 'Q4raw', 'Q5raw',"SEI", 'sc3'])
+    .toFloat();
   return out;
 };
